Jit: Move most x86-64-specific code out of JitCommon
