+++ Source/Core/Core/PowerPC/Jit64/JitRegCache.cpp
@@ -14,7 +14,7 @@
 #include "Common/x64Emitter.h"
 #include "Core/PowerPC/Jit64/Jit.h"
 #include "Core/PowerPC/Jit64/JitRegCache.h"
-#include "Core/PowerPC/JitCommon/Jit_Util.h"
+#include "Core/PowerPC/Jit64Common/Jit64Util.h"
 #include "Core/PowerPC/PowerPC.h"
 
 using namespace Gen;
