+++ Source/Core/Core/PowerPC/Jit64/Jit_SystemRegisters.cpp
@@ -9,7 +9,7 @@
 #include "Core/CoreTiming.h"
 #include "Core/HW/ProcessorInterface.h"
 #include "Core/PowerPC/Jit64/JitRegCache.h"
-#include "Core/PowerPC/JitCommon/Jit_Util.h"
+#include "Core/PowerPC/Jit64Common/Jit64Util.h"
 #include "Core/PowerPC/PowerPC.h"
 
 using namespace Gen;
