+++ Source/Core/Core/PowerPC/Jit64Common/Jit64AsmCommon.cpp
@@ -11,8 +11,8 @@
 #include "Common/x64Emitter.h"
 #include "Core/HW/GPFifo.h"
 #include "Core/PowerPC/Gekko.h"
-#include "Core/PowerPC/JitCommon/JitBase.h"
-#include "Core/PowerPC/JitCommon/Jit_Util.h"
+#include "Core/PowerPC/Jit64Common/Jit64Base.h"
+#include "Core/PowerPC/Jit64Common/Jit64Util.h"
 #include "Core/PowerPC/PowerPC.h"
 
 #define QUANTIZED_REGS_TO_SAVE                                                                     \
